+++ src/components/species-richness-maps/species-richness-maps.vue
@@ -2,11 +2,11 @@
   <div
     class="w-full"
   >
-    <div class="flex justify-between items-end mb-1.5">
+    <div class="flex justify-between items-end">
       <h2 class="text-white text-xl">
-        Number of species detected at each site
+        Distinct species by site
       </h2>
-      <div class="taxon-picker mb-2">
+      <div class="taxon-picker">
         <template
           v-for="a in taxons"
           :key="a.name"
@@ -55,15 +55,13 @@
         </button>
       </div>
     </div>
-    <div
+    <no-data-container-view
       v-if="!hasData"
-      class="text-center text-white"
-    >
-      No data
-    </div>
+      class="h-32 mt-2"
+    />
     <div
       v-show="hasData"
-      class="grid gap-2"
+      class="grid gap-2 mt-2"
       :class="{ [`md:grid-cols-${columnCount}`]: true }"
     >
       <map-bubble-component
@@ -73,9 +71,9 @@
         :taxon="taxon"
         :map-id="idx"
         :map-config="config"
-        class="w-full h-96 "
         :map-style="mapStyle"
         :is-show-labels="isShowLabels"
+        class="w-full"
         @emitMapMoved="mapMoved"
       />
     </div>